Fiddlin' with the testmap and more.. In the next episode of [A Crossover Episode]. Airing Thursdays at 9PM. Stay Tuned.
Updated Emotion to 289.
Fixed the Wanderer not getting the prop InteractScript from the json file.
Tweaked some values.
Tried to make the dialog box work. It works. (Was already working.)
Added a code snippet to enable the player to manually close the dialog box when not in range of any targets.